refactor(schema): split parser.rs into submodules
The schema parser was 1415 LOC - 415 over the repo's own <1000 LOC
file budget. The Python and Go sibling ports keep their parsers at
705 and 760 LOC respectively; this brings the Rust port into line.

Split src/schema/parser.rs -> src/schema/parser/ with one submodule
per DEFINE statement class plus a parent mod.rs holding the shared
JSON/regex helpers and DatabaseInfo:

  mod.rs   745 - public re-exports, DatabaseInfo, shared helpers,
                 existing 49-test suite (unchanged).
  field.rs 196 - DEFINE FIELD parser + type/clause extractors.
  index.rs 220 - DEFINE INDEX parser (UNIQUE/SEARCH/MTREE/HNSW).
  event.rs  67 - DEFINE EVENT parser.
  access.rs143 - DEFINE ACCESS parser (JWT + RECORD).
  table.rs  78 - parse_table_mode + parse_table_info.
  db.rs    107 - parse_db_info + relation-endpoint extractor.

Pure refactor: no behaviour change, no API change. Every public symbol
(parse_field, parse_fields, parse_index, parse_indexes, parse_event,
parse_events, parse_access, parse_table_mode, parse_table_info,
parse_db_info, DatabaseInfo) still resolves at crate::schema::parser::
with identical signatures. The test module was kept intact against the
new mod.rs; extract_relation_endpoints is pub(super) to stay reachable
for the two tests that exercise it directly.

Validation:
  cargo fmt --all --check               passes
  cargo clippy --all-targets --all-features -- -D warnings passes
  cargo test --lib --all-features       1030 passed (49 parser)
  cargo test --doc --all-features       73 passed
  cargo test --test '*' --all-features --test-threads=1
                                        55 passed against
                                        surrealdb/surrealdb:v3.0.5

Refs #72, umbrella #49.
